Add namespace Fog, make bind-gen ignore fog-namespace
The wrapped methods used to be included in libfog.lib as e.g. draw() and
update() directly, making games written in C++ unable to name a function
draw(), update() and so on. With this they're reachable with a
`fog_`-prefix when compiling with C and the `Fog::`-namespace when
compiling with C++, which is more reasonable.

The change in bind-gen is to avoid functions named stuff like
`fog_fog_run`.